pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "interact-morph"
version = "0.1.0"
description = "INTERACT-Morph data and modeling foundation"
readme = "README.md"
requires-python = ">=3.10"
dependencies = []

[project.scripts]
interact-morph = "interact_morph.cli:main"

[project.optional-dependencies]
validation = [
  "jsonschema>=4.0",
  "PyYAML>=6.0"
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
